merge, radix e bucket ainda sem troca

+++ bolha.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include <time.h>
 
-void bolha(int *vetor, int n, long int *comp, long int *troca)
+void bolha(int *vetor, int n, long long int *comp, long long int *troca)
 {
     int i, j = 0, aux;
     for ( i = n-1; i > 0; i--)
@@ -24,11 +24,11 @@ void chamaBolha(int *vetor, int n)
 {
     clock_t start, end;
     start = clock();
-    long int comp = 0;
-    long int troca = 0;
+    long long int comp = 0;
+    long long int troca = 0;
 
     bolha(vetor, n, &comp, &troca);
 
     end = clock();
-    printf("\nBolha com %d elementos\nQuantidade de Comparacoes: %ld\nQuantidade de Trocas: %ld\nTempo de execucao: %f segundos\n\n", n, comp, troca, ((double)(end - start)) / CLOCKS_PER_SEC);
+    printf("\nBolha com %d elementos\nQuantidade de Comparacoes: %lld\nQuantidade de Trocas: %lld\nTempo de execucao: %f segundos\n\n", n, comp, troca, ((double)(end - start)) / CLOCKS_PER_SEC);
 }
