+++ bolha.h
@@ -1,2 +1,2 @@
-void bolha(int *vetor, int n, long int *comp, long int *troca);
+void bolha(int *vetor, int n, long long int *comp, long long int *troca);
 void chamaBolha(int *vetor, int n);
