+++ bolhaComParada.c
@@ -6,8 +6,8 @@ void bolhaComParada(int *vetor, int n)
 {
     clock_t start, end;
     start = clock();
-    long int comp = 0;
-    long int troca = 0;
+    long long int comp = 0;
+    long long int troca = 0;
     int aux = 0;
     int ultimoIndex = n - 1;
     bool trocou = true;
@@ -32,5 +32,5 @@ void bolhaComParada(int *vetor, int n)
         ultimoIndex = novoUltimoIndex;
     }
     end = clock();
-    printf("\nBolha(com parada) com %d elementos\nQuantidade de Comparacoes: %ld\nQuantidade de Trocas: %ld\nTempo de execucao: %f segundos\n\n", n, comp, troca, ((double)(end - start)) / CLOCKS_PER_SEC);
+    printf("\nBolha(com parada) com %d elementos\nQuantidade de Comparacoes: %lld\nQuantidade de Trocas: %lld\nTempo de execucao: %f segundos\n\n", n, comp, troca, ((double)(end - start)) / CLOCKS_PER_SEC);
 }
