+++ insercaoBinaria.c
@@ -6,8 +6,8 @@ void insercaoBinaria(int *vet, int n)
     clock_t start, end;
     start = clock();
 
-    long int comp = 0;
-    long int troca = 0;
+    long long int comp = 0;
+    long long int troca = 0;
     int aux = 0;
     int dir = 0, esq = 0, meio = 0;
 
@@ -41,5 +41,5 @@ void insercaoBinaria(int *vet, int n)
     }
 
     end = clock();
-    printf("\nInsercao Binaria com %d elementos\nQuantidade de Comparacoes: %ld\nQuantidade de Trocas: %ld\nTempo de execucao: %f segundos\n\n", n, comp, troca, ((double)(end - start)) / CLOCKS_PER_SEC);
+    printf("\nInsercao Binaria com %d elementos\nQuantidade de Comparacoes: %lld\nQuantidade de Trocas: %lld\nTempo de execucao: %f segundos\n\n", n, comp, troca, ((double)(end - start)) / CLOCKS_PER_SEC);
 }
