+++ insercaoDireta.c
@@ -4,8 +4,8 @@
 void insercaoDireta(int *vetor, int n){
     clock_t start, end;
     start = clock();
-    long int comp = 0;
-    long int troca = 0;
+    long long int comp = 0;
+    long long int troca = 0;
     int i, auxiliar, j;
     for(i = 1; i < n; i++)
     {
@@ -31,5 +31,5 @@ void insercaoDireta(int *vetor, int n){
         }
     }
     end = clock();
-    printf("\nInsercao Direta com %d elementos\nQuantidade de Comparacoes: %ld\nQuantidade de Trocas: %ld\nTempo de execucao: %f segundos\n\n", n, comp, troca, ((double)(end - start)) / CLOCKS_PER_SEC);
+    printf("\nInsercao Direta com %d elementos\nQuantidade de Comparacoes: %lld\nQuantidade de Trocas: %lld\nTempo de execucao: %f segundos\n\n", n, comp, troca, ((double)(end - start)) / CLOCKS_PER_SEC);
 }
