+++ shellSort.c
@@ -6,8 +6,8 @@ void shellsort(int *vetor, int n)
     clock_t start, end;
     start = clock();
     
-    long int troca = 0;
-    long int comp = 0;
+    long long int troca = 0;
+    long long int comp = 0;
     int i, j, h = 1, aux;
     while (h < n)
     {
@@ -40,5 +40,5 @@ void shellsort(int *vetor, int n)
         }
     }
     end = clock();
-    printf("\nShellsort com %d elementos\nQuantidade de Comparacoes: %ld\nQuantidade de Trocas: %ld\nTempo de execucao: %f segundos\n\n", n, comp, troca, ((double)(end - start)) / CLOCKS_PER_SEC);
+    printf("\nShellsort com %d elementos\nQuantidade de Comparacoes: %lld\nQuantidade de Trocas: %lld\nTempo de execucao: %f segundos\n\n", n, comp, troca, ((double)(end - start)) / CLOCKS_PER_SEC);
 }
